Reference

Here you can find the documentation for the all the functions and structs in the library.

QuTaM.DetectionClick — Type

DetectionClick(time::Float64, label::Int64)

Inmutable struct that represents the clicks by the time waited to see the click and the label of the channel in which it occured.

Fields

time::Float64: Waiting time

label::Int64: Label of the channel of the click

QuTaM.SimulParameters — Type

Inner constructor of SimulParameters SimulParameters(psi0::Vector{ComplexF64}, tf::Float64, s::Int64, ntraj::Int64, nsamples::Int64=10000, m::Float64=10.0, eps::Float64=1e-3)

QuTaM.SimulParameters — Type

SimulParameters(
    psi0::Vector{ComplexF64}, nsamples::Int64, seed::Int64,
            ntraj::Int64, multiplier::Float64, tf::Float64,
            dt::Float64, eps::Float64)

A mutable struct containing all the necessary information for running the the simulation.

Fields

psi0::Vector{ComplexF64}: Initial state vector

nsamples::Int64: Number of samples in the finegrid

seed::Int64: seed

ntraj::Int64: Number of trajectories

multiplier::Float64: Multiplier to use in the fine grid

tf::Float64: Final time

dt::Float64: time step for the finegrid

eps::Float64: Tolerance for passing WTD normalziation

Constructor

To create an instance it's enough to provide initial state, final time, seed and number of trajectories. Unless given nsamples, multiplier and eps use default values. SimulParameters(psi0::Vector{ComplexF64}, tf::Float64, s::Int64, ntraj::Int64, nsamples::Int64=10000, m::Float64=10.0, eps::Float64=1e-3)

About the multiplier

For the Gillipsie algorithm to work it's key to have a grid that's capable of resolving the statistical details of the WTD, this grid is taken in the interval (0, tf*multiplier).

QuTaM.System — Type

System(
NLEVELS::Int64, NCHANNELS::Int64, H::Matrix{ComplexF64}
Ls::Vector{Matrix{ComplexF64}}, J::Matrix{ComplexF64},
Heff::Matrix{ComplexF64})

A mutable struct that characterizes the dynamics via specification of the jump and hamiltonian operators.

Fields

NLEVELS::Int64: Number of levels of the system

NCHANNELS::Int64: Number of jump channels

H::Matrix{ComplexF64}: Hamiltonian

Ls::Vector{Matrix{ComplexF64}}: List of jump operators

J::Matrix{ComplexF64}: Sum of all the $L_k^{*}L_k$

Heff::Matrix{ComplexF64}: Effective Hamiltonian

Constructor

To create an instance it's enough to provide the hamiltonian and the jump operators in a vector. System(H::Matrix{ComplexF64}, Ls::Vector{Matrix{ComplexF64}})

QuTaM.System — Method

     Inner Constructor of `System` struct.
     # Arguments:
     `H::Matrix{ComplexF64}`
     `Ls::Vector{Matrix{ComplexF64}}`

QuTaM.Trajectory — Type

Alias for Vector{DetectionClick}

QuTaM.precompute! — Method

precompute!(sys::System, nsamples::Int64,
     ts::Vector{Float64}, Qs::Vector{Matrix{ComplexF64}})

Does the precomputation routine for the Gillipsie algorithm. The result is stored in the Qs.

Arguments

sys::System: system nsamples::Int64: number of samples ts::Vector{Float64}: fine grid vector Qs::Vector{Matrix{ComplexF64}}: vector of matrices to store the precomputation.

QuTaM.run_single_trajectory — Method

run_single_trajectories(sys::System, params::SimulParameters,
 W::Vector{Float64}, P::Vector{Float64}, psi::Vector{ComplexF64},
 ts::Vector{Float64}, Qs::Vector{Matrix{ComplexF64}};
  seed::Int64 = 1) -> Trajectory

Sample a single trajectory from the system and parameters using the Gillipsie algorithm . This is inteded to be used by run_trajectories

Arguments:

sys::System: System of interest

params::SimulParameters: simulation parameters

W::Vector{Float64}: to store the weights over the fine grid

P::Vector{Float64}: to store the weights over the channels

psi::Vector{ComplexF64}: to store the current state vector

ts::Vector{Float64}: the finegrid of waiting times

Qs::Vector{Matrix{ComplexF64}}: to store the precomputed values

seed::Int64: seed for generating the trajectory

Returns:

The sample trajectory

Warning: the seed does not coincide with that of params by default.

Waning: final jump in the trajectory happens after final time

The trajectory ends when a jump that happens after params.tf is obtained, yet that jump is stored in the trajectory. In other words, the last jump of the trajectory always happen after the set final time.

QuTaM.run_trajectories — Method

run_trajectories(sys::System, params::SimulParameters) -> Vector{Trajectory}

Sample multiple trajectories for a given system and parameters.

Arguments

sys::System: The quantum system to simulate, containing information about its structure, energy levels, and dynamics.

params::SimulParameters: A structure containing simulation parameters such as:

Returns

Vector{Trajectory}: A vector containing the results of the simulated trajectories. Each element corresponds to a single trajectory and encapsulates relevant system state information over time.

QuTaM.sample_single_trajectory — Method

sample_single_trajectory(sys::System, params::SimulParameters, seed::Int64) -> Vector{Trajectory}

Sample a single trajectory. This is inteded to be used in case a single sample is needed without redefining params.ntraj.

Arguments

sys::System

params::SimulParameters

seed::Int64

Returnrs

A 1-element trajectory vector.

QuTaM.states_at_jumps — Method

states_at_jumps(traj::Trajectory, sys::System,
                  psi0::Vector{ComplexF64}) - > Array{ComplexF64}

From a given trajectory, recover the states at each jump.

Arguments

traj::Trajectory: Trajectory

sys::System: System of interest

psi0::Vector{ComplexF64}: Initial state

Returns

Array{ComplexF64} with the states

The dimensions of the returned array s are (size(traj), sys.NLEVELS), so to recover the state vector at the $n$-th jump one would do s[n, :].