Pauli Matrices

BackAction.sigma_x — Constant

Pauli Matrix $\sigma_x = |0\rangle\langle 1| + |1\rangle\langle 0|$

BackAction.sigma_y — Constant

Pauli Matrix $\sigma_y = i|0\rangle\langle 1| - i|1\rangle\langle 0|$

BackAction.sigma_z — Constant

Pauli Matrix $\sigma_z = |0\rangle\langle 0| - |1\rangle\langle 1|$

BackAction.sigma_m — Constant

Lowering operator $\sigma_- = |0\rangle\langle 1|$

BackAction.sigma_p — Constant

Raising operator $\sigma_+ = |1\rangle\langle 0|$

Runge-Kutta 4 Solver

To run the test against the Lindblad equation, the library includes a simple implementation of the Runge-Kutta4 algorithm for ODEs.

BackAction.rk4 — Function

rk4(f, y0, tspan, nsteps)

Arguments

f::Function: function $f(t, y)$ defininf the system of ODEs

y0::Vector{Float}: inital condition

tspan::Tuple{Float64}: initial and final time

nsteps::Int: number of steps.

Returns

An array Array{typeof(y0[1])} with dimensions (ndims(y0), nsteps) with the values of $y$ at each step.